Finish system report with table overview.
